services/django-backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
python-decouple==3.8
dj-database-url==2.2.0
gunicorn==22.0.0
uvicorn[standard]==0.30.1
whitenoise==6.7.0
requests==2.32.3
pydantic==2.8.2
